+++ src/Helpers/JsHelpers.js
@@ -1,3 +1,10 @@
+export class JsHelpers {
+    static addSecondsToDate(date, seconds) {
+        return new Date(date.getTime() + seconds * 1000);
+    }
+}
+
+
 export function dateToStringHhMmSs(date) {
     return alwaysTwoDigits(date.getHours()) + ":" + alwaysTwoDigits(date.getMinutes()) + ":" + alwaysTwoDigits(date.getSeconds());
 }